Finish styling on the authentication modals
