--- src/containers/Admin/LoginModal/LoginOptions/styles.scss
@@ -1,1 +1,0 @@
-@import '../../../../assets/styles/import.scss';
