--- src/containers/Admin/LoginModal/ResetPassword/styles.scss
@@ -1,5 +1,0 @@
-@import '../../../../assets/styles/import.scss';
-
-.go-to:hover {
-    cursor: pointer;
-}
